Lots of fixes
Fixed host and connect button in startmenu. Fixed generatorcontroller so that generatorUI isn't null if the game is already inactive when you join a server.
Fixed that OxygenHandler was calling a server only method on all the clients using ClientRPC.
Fixed the rounding in displayfloatnbr and made it more dynamic (you can round to any amount of decimals instead of specifically 2)