Big changes
Merged bullet scripts so we have one unified script for both enemy and allied bullets. Added new cannon stations. Merged all the cannon scripts into a single script that works for all cannons. Fixed the submarine prefab to work with the new cannon script. Fixed bullets so they go the right way for the submarine. Added a no collision period for bullets so they don't instantly self destruct if they collide with their creators when instantiated.

Need to test whether bullets collide correctly with the submarine and the enemy ships.